requirements.txt
automium==0.2.6
automium-web==0.1.1
Click==7.0
Django==3.0.2
django-model-utils==3.2.0
djangorestframework==3.10.3
ez-setup==0.9
Flask==1.1.1
freeze==1.0.10
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
MarkupSafe==1.1.1
mysqlclient==1.4.6
psycopg2==2.8.4
pytz==2019.3
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
views==0.3
Werkzeug==0.16.0
XlsxWriter==1.2.7
